+++ src/main/java/com/example/assignmentgd1/model/HoaDonChiTiet.java
@@ -1,2 +1,36 @@
-package com.example.assignmentgd1.model;public class HoaDonChiTiet {
+package com.example.assignmentgd1.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "hoa_don_chi_tiet")
+public class HoaDonChiTiet {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @ManyToOne
+    @JoinColumn(name = "id_hd")
+    private HoaDon hoaDon;
+
+    @ManyToOne
+    @JoinColumn(name = "id_spct")
+    private SanPhamChiTiet sanPhamChiTiet;
+
+    private Integer soLuong;
+
+    private Double donGia;
+
+    private Boolean trangThai;
+
 }
+
