--- src/main/java/com/example/assignmentgd1/model/HoaDonChiTiet.java~
@@ -1,42 +1,0 @@
-package com.example.assignmentgd1.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "hoa_don_chi_tiet")
-public class HoaDonChiTiet {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @ManyToOne
-    @JoinColumn(name = "id_hd")
-    private HoaDon hoaDon;
-
-    @ManyToOne
-    @JoinColumn(name = "id_spct")
-    private SanPhamChiTiet sanPhamChiTiet;
-
-    private Integer soLuong;
-
-    private Double donGia;
-
-    private Boolean trangThai;
-
-    List<HoaDonChiTiet> dsHdct; // <--->
-
-}
