+++ src/main/resources/application.properties
@@ -8,6 +8,5 @@ spring.datasource.password=123456
 spring.datasource.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
 
 spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.show-sql=true
