--- src/main/resources/application.properties~
@@ -1,13 +1,0 @@
-spring.application.name=AssignmentGD1
-spring.mvc.view.prefix=/WEB-INF/views/
-spring.mvc.view.suffix=.jsp
-
-spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=ASM_Java5;encrypt=false
-spring.datasource.username=sa
-spring.datasource.password=123456
-spring.datasource.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
-
-#spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-
-
