+++ src/test/java/com/example/assignmentgd1/AssignmentGd1ApplicationTest.java
@@ -1,4 +1,10 @@
-import static org.junit.jupiter.api.Assertions.*;
+package com.example.assignmentgd1;
+
+import org.junit.jupiter.api.Test;
+
 class AssignmentGd1ApplicationTest {
-  
+
+    @Test
+    void main() {
+    }
 }
